Workaround for NaN while computing Histogram.uniform
The dataset Public_Schools_Oregon.csv is causing a NaN for the p
parameter received by Histogram.extendedSum() while computing
Histogram.createBinSumMap().  The latter runs a loop ignoring
SumOutOfRangeException, so i'm wildly guessing that p can be NaN and
should be ignored.

While this is probably wrong, at least avoids a null pointer
exception (when p is NaN in extendedSum(p) that happens because
_bins.higherEntry(p) in Histogram.java:331 return null).